# worker-app/docs/api/mobile.openapi.yaml
openapi: 3.0.3
info:
  title: Shveyka Worker Mobile API
  version: 1.0.0
  description: >
    Mobile worker API for shop-floor execution, task handling, cutting nastils,
    and entry capture.
servers:
  - url: /api
...
paths:
...
  /mobile/tasks/{id}/entries:
    get:
      tags: [Entries]
      summary: List task entries for a worker task
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Entry list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskEntry"
    post:
      tags: [Entries]
      summary: Create a task entry
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskEntryCreateRequest"
      responses:
        "200":
          description: Entry stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskEntry"
...
  /mobile/batches/{id}/route-card:
    get:
      tags: [Batches]
      summary: Get the route card for a batch
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Route card payload.
...
components:
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    TaskEntryCreateRequest:
      type: object
      required: [batch_id, operation_id]
      properties:
        batch_id:
          type: integer
        operation_id:
          type: integer
        quantity:
          type: integer
          description: Generic quantity payload for simple operations
          nullable: true
        defect_quantity:
          type: integer
          default: 0
          description: Generic defect quantity for simple operations
        metric_value:
          type: number
          default: 0
        size:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        local_id:
          type: string
          nullable: true
        status:
          type: string
        size_rows:
          type: array
          description: Overlock payload. Each row is bounded by the cutting breakdown for the same size.
          items:
            type: object
            required: [size, quantity, defect_quantity]
            properties:
              size:
                type: string
              quantity:
                type: integer
                minimum: 0
              defect_quantity:
                type: integer
                minimum: 0
    TaskEntry:
      type: object
      additionalProperties: true
...